Feature: Model for statistics (#205)
* Add tests for creating audit events

Adds test for creating audit events to existing test infrastructure.

* Introduce model for audit events

Adds the model, migration and factory and updates the schema. Additionally, it introduces a "has_many" connection to the item model. It destroys the item with the audit events, because they won't be relevant anymore.

* Add audit event helper

Introduces the audit event helper as interface for audit event. User are aware of available audit event types now.

* Use audit events in existing infrastructure

Adds the creation of audit events to the existing infrastructure (mostly in item controller). Therefore, I had to disable the rubocop abc-metric for a function, because I couldn't reduce the complexity.

* Add empty statistics helper

Introduces empty statistics helper to add functions for calculating statistics and metrices in the future.

Co-authored-by: Lietze <50149435+Lietze@users.noreply.github.com>
Co-authored-by: Luke <37272151+SaturnHafen@users.noreply.github.com>